move sigchain.go to por/ from sigchain/
move sigchain.go to por/ from sigchain and rename sigchain/ to sigchains/.
